+++ CHANGES.rst
@@ -7,7 +7,7 @@ Changelog
 - Added upgrade step: put our control panel action in the Products
   (addons) category.  And give it a proper icon (gif on Plone 3, png
   on Plone 4).  Set the default on install to png so it is best for
-  Plone 4.
+  Plone 4.  Same for the tool icon in portal_types.
   [maurits]
 
 - Fixed javascript error for undefined tinymce due to typo.  Note that